tests/test_requirements.txt
-e .
flake8
requests==2.7.0
django-haystack==2.5.1
pysolr==3.3.2
python-dateutil==2.4.2
django-rq==0.9.1
boto==2.38.0
boto3==1.9.22
pytest==3.7.0
pytest-django==3.3.3